refactor(storage): move SqliteSnapshot to shared Mail2Pst.Core.Storage
